# Reduce excessive memory allocations in parseNumber and formatNumber tags