--- a/impl/src/main/java/org/apache/taglibs/standard/tag/common/fmt/FormatNumberSupport.java
+++ b/impl/src/main/java/org/apache/taglibs/standard/tag/common/fmt/FormatNumberSupport.java
@@ -172,7 +172,7 @@ public abstract class FormatNumberSupport extends BodyTagSupport {
                 pageContext,
                 this,
                 true,
-                NumberFormat.getAvailableLocales());
+                SetLocaleSupport.numberLocales);
 
         if (loc != null) {
             // Create formatter