During profiling it appeared that _formatNumber_ tag  causes huge memory pressure which overwhelms GC resulting in overall application slowness.
![before](https://user-images.githubusercontent.com/722112/106400125-58202000-6425-11eb-9317-cf11c18e1780.png) 
|.| Before | After |
|-----------| ----------- |----------- |
| RPS |  578.0/s   |   707.7/s  |
| Avg Latency |  692 ms   |  561 ms  |